Update to divide the components
