+++ server/src/main/java/com/albercafe/rabbitmarket/controller/AuthController.java
@@ -16,6 +16,7 @@ import javax.validation.Valid;
 @RestController
 @RequestMapping("/api/auth")
 @AllArgsConstructor
+@CrossOrigin(origins = "http://localhost:3000/#")
 public class AuthController {
 
     private final AuthService authService;
