+++ server/src/main/resources/.gitignore
@@ -1,0 +1,2 @@
+application.properties
+rabbit-market.jks
