+++ server/src/main/resources/application.properties
@@ -7,13 +7,14 @@ spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5InnoDBDialec
 spring.jpa.hibernate.ddl-auto=update
 
 # Mail
-spring.mail.host=smtp.mailtrap.io
-spring.mail.port=2525
-spring.mail.username=fadac19c891a3d
-spring.mail.password=adb294dd1ce59b
-spring.mail.protocol=smtp
+spring.mail.host=smtp.gmail.com
+spring.mail.port=587
+spring.mail.username=veloci1024@gmail.com
+spring.mail.password=vel1024cel512
+spring.mail.properties.mail.smtp.auth=true
+spring.mail.properties.mail.smtp.starttls.enable=true
 
 # keystore
 jks.password=rabbit1234
-jwt.expiration.time=90000
+jwt.expiration.time=900000
 # ms 단위 15분
